scripts/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
mediapipe==0.10.7
pillow==10.1.0
numpy==1.24.3
